backend/requirements.txt
Flask==3.1.2
xarray>=2024.7.0
cfgrib>=0.9.15.0
numpy>=2.0.0
Pillow>=10.4.0
eccodes>=2.43.0
Flask-Cors>=4.0.0
